maxPool+Dense, 82acc, (categorical_crossentropy)
This is the final model that with a few additional dense layer and MaxPooling setup. Also there is a bug on using sparse_categorical_crossentropy: The pretrained (saved) model doesn't show correct accuracy values on load_model. When you change the loss function to categorical_crossentropy, it's solved. This model has 82% accuracy with 0.520 loss value.